requirements.txt
Flask 
Flask_login
python-docx
 